lvgl: remove FontAwesome font and add a license file for Montserrat font to prevent licensing issues
In the current configuration of LVGL library, a font
at 14 pt is available by default. This font consists
of Montserrat and FontAwesome. Montserrat is licensed
under SIL Open Font License, but license of FontAwesome
is unclear. To prevent licensing issues, this commit
removes FontAwesome font and adds a license file for
Montserrat font.
The addition of the license file was suggested at the
following issue. It is still not closed.
  https://github.com/lvgl/lvgl/issues/6711

Signed-off-by: Yoshihito Ogawa <yoshihito.ogawa.kc@renesas.com>